added testcase cc transmission

+++ assignment4/test_meltdown-template.c
@@ -31,7 +31,7 @@ void test_cc_transmission(){
     for(int i=0; i<256;i++){
         cc_setup();
         cc_transmit(i);
-        uint64_t time = cc_receive();
+        int time = cc_receive();
         CU_ASSERT_EQUAL(time, i);
         if (time !=i){
             printf("transmission failed at %i\n", i);
